+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog for Alchemy SDK for JavaScript
 
 ## Unreleased
+- Added a `size` field to the `Media` object in the NFT metadata responses to indicate the size of the media in bytes.
+
+## 2.0.3
 
 ### Minor Changes
 
@@ -8,7 +11,8 @@
 - Added support for a Wallet object. This object is identical to the Ethers `Wallet` class.
 - Added support for Arbitrum Goerli network via the `Network.ARB_GOERLI` enum.
 - Added support for the AStar Mainnet network via the `Network.ASTAR_MAINNET` enum.
-- Added a `size` field to the `Media` object in the NFT metadata responses to indicate the size of the media in bytes.
+- Added typed request/response overloads for `CoreNamespace.getAssetTransfers()` for when the `withMetadata` param is true.
+- Added support the `tokenUriTimeoutInMs` parameter on `getNftsForOwner()`, `getNftMetadata()`, and `getNftsForContract()`.
 
 ## 2.0.2